Move basePathLink to shared utils
